fix ema
